requirements.txt
torch
torchvision
kaggle
Pillow
matplotlib
numpy
pandas
scikit-learn